rename retrofit package to network
